Fix bug in html2latex converter (locking converted parts).
When inserting and locking LaTeX into the string which is converted, it may loop when the key generation is not unique enough.
By default the length of the key is the same as text which is locked. The key has now a length of at least 10 characters for preventing the key generation from looping.