Listing converter: avoid using beatiful soup.
When using beautiful soup, tags may be changed, causing other patterns
to no longer match in the HTML contained in the list items.

Example:
  <li>Foo<keyword title="foo" /> Bar</li>
will be converted into:
  <li>Foo<keyword title="foo"> Bar</keyword></li>
which will no longer be matched by the keyword subconverter (ftw.book).

To avoid such problems we only use BeatifulSoup when the HTML is broken
and cannot be parsed by the xml parser.
In this case we must put up with no longer working subconverters.