Fix encoding error by passing charset hint to BeautifulSoup when converting tables:
Non-ASCII characters that *only* appear in table headings seem to
trip up BeautifulSoups charset sniffing. It's therefore important
to pass a character set hint to BeautifulSoup, so it doesn't try
to sniff the content / do any flawed charset detection that leads
to Mojibake in the final LaTeX output.

This is especially true because we usually control the HTML input
that ends up in ftw.pdfgenerator, so we don't really have a need
for BeautifulSoup's charset sniffing.