Fix underscore escaping in hyperlink label.
The label part of a \href{url}{label} type hyperlink was not escaped
correctly regarding underscores.

The underscores were escaped with two backslashes (\\_).
This escapes the slash, not the underscore. Which has the effect that
an additional slash is added and the underscore is not escaped
anymore.

This causes errors such as "! Missing $ inserted.".

The fix is to only escape with one slash (\_).